Brought AROS grub2 in sync with grub2 repository revision 1919
git-svn-id: https://svn.aros.org/svn/aros/trunk@30086 fb15a70f-31f2-0310-bbcc-cdcc74a49acc

+++ arch/common/boot/grub2/commands/blocklist.c
@@ -72,7 +72,7 @@ grub_cmd_blocklist (struct grub_arg_list *state __attribute__ ((unused)),
       if (num_entries++)
 	grub_printf (",");
 
-      grub_printf ("%llu", sector - part_start);
+      grub_printf ("%llu", (unsigned long long) (sector - part_start));
       if (num > 0)
 	grub_printf ("+%u", num);
       if (offset != 0 || length != 0)
